+++ src/Components/Login/login.css
@@ -17,7 +17,12 @@
   justify-content: center;
   align-items: center;
   overflow: hidden;
+  transform: scale(0);
   transition: transform 0.5s ease, height 0.2s ease;
+}
+
+.wrapper.active-popup {
+  transform: scale(1);
 }
 
 .wrapper.active {
@@ -176,30 +181,4 @@
 
 .login-register p a:hover {
   text-decoration: underline;
-}
-
-.WrongCredentials {
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  width: 100%;
-  text-align: center;
-  padding: 1rem 0 0 0;
-}
-
-.WrongCredentials p {
-  color: red;
-  font-family: math;
-  width: 75%;
-  font-size: 16px;
-  gap: 0.3rem;
-}
-
-.WrongCredentials img {
-  width: 6%;
-}
-
-#wrongCredentials {
-  display: none;
-  font-family: 'Courier New', Courier, monospace;
 }
